osmflat: add criterion bench and test_support helpers
