+++ osmflat/src/lib.rs
@@ -41,6 +41,9 @@ include!("osmflat_generated.rs");
 mod spatial;
 mod tags;
 
+#[cfg(any(test, feature = "test-support"))]
+pub mod test_support;
+
 pub use crate::osm::*;
 pub use crate::spatial::*;
 pub use crate::tags::*;
