+++ osmflat/src/spatial.rs
@@ -256,134 +256,7 @@ pub fn find_relations_by_bounding_box(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{Header, Node, NodeIndex, Osm, OsmBuilder, Relation, Way};
-    use flatdata::MemoryResourceStorage;
-
-    const COORD_SCALE: i32 = 10_000_000;
-
-    fn scale(v: f64) -> i32 {
-        (v * COORD_SCALE as f64) as i32
-    }
-
-    fn unscale(v: i32) -> f64 {
-        v as f64 / COORD_SCALE as f64
-    }
-
-    /// Build an in-memory archive laid out exactly the way `osmflatc` does:
-    /// nodes ordered by the z-order curve, ways and relations ordered by the
-    /// `XZ2SFC` index of their bounding box, each vector terminated by a
-    /// sentinel. This lets the query functions binary-search correctly.
-    ///
-    /// - `node_lonlat`: `(lon, lat)` per node, in degrees.
-    /// - `ways`: each way as a list of indices into `node_lonlat`.
-    /// - `relation_bboxes`: each relation as `Some((min_lon, min_lat, max_lon,
-    ///   max_lat))`, or `None` for a no-location (sentinel) relation.
-    fn build_archive(
-        node_lonlat: &[(f64, f64)],
-        ways: &[Vec<usize>],
-        relation_bboxes: &[Option<(f64, f64, f64, f64)>],
-    ) -> Osm {
-        let storage = MemoryResourceStorage::new("/test");
-        let builder = OsmBuilder::new(storage.clone()).unwrap();
-
-        let mut header = Header::new();
-        header.set_coord_scale(COORD_SCALE);
-        builder.set_header(&header).unwrap();
-
-        // Nodes, ordered by the z-order curve.
-        let curve = node_curve();
-        let mut order: Vec<usize> = (0..node_lonlat.len()).collect();
-        order.sort_by_key(|&i| node_index(&curve, node_lonlat[i].0, node_lonlat[i].1));
-        let mut final_idx = vec![0u64; node_lonlat.len()];
-        for (pos, &orig) in order.iter().enumerate() {
-            final_idx[orig] = pos as u64;
-        }
-        let mut node_vec: Vec<Node> = order
-            .iter()
-            .map(|&orig| {
-                let (lon, lat) = node_lonlat[orig];
-                let mut n = unsafe { Node::new_unchecked() };
-                n.set_lon(scale(lon));
-                n.set_lat(scale(lat));
-                n.set_tag_first_idx(0);
-                n
-            })
-            .collect();
-        node_vec.push(unsafe { Node::new_unchecked() }); // sentinel
-        builder.set_nodes(&node_vec).unwrap();
-
-        // Ways, ordered by the bounding-box curve.
-        let wcurve = way_curve();
-        let way_bbox = |w: &[usize]| -> (f64, f64, f64, f64) {
-            let mut bb: Option<(f64, f64, f64, f64)> = None;
-            for &ni in w {
-                let (lon, lat) = node_lonlat[ni];
-                bb = Some(match bb {
-                    Some((a, b, c, d)) => (a.min(lon), b.min(lat), c.max(lon), d.max(lat)),
-                    None => (lon, lat, lon, lat),
-                });
-            }
-            bb.unwrap()
-        };
-        let mut worder: Vec<usize> = (0..ways.len()).collect();
-        worder.sort_by_key(|&i| {
-            let (a, b, c, d) = way_bbox(&ways[i]);
-            bbox_index(&wcurve, a, b, c, d)
-        });
-        let mut way_vec: Vec<Way> = Vec::new();
-        let mut nodes_index_vec: Vec<NodeIndex> = Vec::new();
-        for &wi in &worder {
-            let mut w = unsafe { Way::new_unchecked() };
-            w.set_tag_first_idx(0);
-            w.set_ref_first_idx(nodes_index_vec.len() as u64);
-            for &ni in &ways[wi] {
-                let mut idx = NodeIndex::new();
-                idx.set_value(Some(final_idx[ni]));
-                nodes_index_vec.push(idx);
-            }
-            way_vec.push(w);
-        }
-        let mut sentinel_way = unsafe { Way::new_unchecked() }; // sentinel terminates the ref range
-        sentinel_way.set_tag_first_idx(0);
-        sentinel_way.set_ref_first_idx(nodes_index_vec.len() as u64);
-        way_vec.push(sentinel_way);
-        builder.set_ways(&way_vec).unwrap();
-        builder.set_nodes_index(&nodes_index_vec).unwrap();
-
-        // Relations, ordered by the bounding-box curve; no-location relations
-        // get the sentinel mbb and sort last (key u64::MAX).
-        let rel_key = |bbox: Option<(f64, f64, f64, f64)>| match bbox {
-            Some((a, b, c, d)) => bbox_index(&wcurve, a, b, c, d),
-            None => u64::MAX,
-        };
-        let mut rorder: Vec<usize> = (0..relation_bboxes.len()).collect();
-        rorder.sort_by_key(|&i| rel_key(relation_bboxes[i]));
-        let mut rel_vec: Vec<Relation> = rorder
-            .iter()
-            .map(|&ri| {
-                let mbb = relation_bboxes[ri]
-                    .map(|(a, b, c, d)| [scale(a), scale(b), scale(c), scale(d)])
-                    .unwrap_or(RELATION_NO_BBOX);
-                let mut r = unsafe { Relation::new_unchecked() };
-                r.set_tag_first_idx(0);
-                r.set_min_lon(mbb[0]);
-                r.set_min_lat(mbb[1]);
-                r.set_max_lon(mbb[2]);
-                r.set_max_lat(mbb[3]);
-                r
-            })
-            .collect();
-        rel_vec.push(unsafe { Relation::new_unchecked() }); // sentinel
-        builder.set_relations(&rel_vec).unwrap();
-
-        // Remaining resources are required to open the archive but unused here.
-        builder.start_relation_members().unwrap().close().unwrap();
-        builder.set_tags(&[]).unwrap();
-        builder.set_tags_index(&[]).unwrap();
-        builder.set_stringtable(b"\0").unwrap();
-
-        Osm::open(storage).unwrap()
-    }
+    use crate::test_support::{build_archive, unscale};
 
     #[test]
     fn finds_nodes_in_box_and_excludes_outside() {
